requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
SQLAlchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.3
passlib[bcrypt]==1.7.4
redis==5.0.8
